requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
tensorflow>=2.13.0
keras>=2.13.0
gunicorn>=21.2.0